+++ eslint.config.js
@@ -1,5 +1,5 @@
 // @ts-check
-// eslint.config.js 2.0.3
+// eslint.config.js 2.0.5
 
 // This ESLint configuration is designed for a TypeScript project using ESM modules.
 
@@ -19,29 +19,15 @@ import importsort from 'eslint-plugin-simple-import-sort';
 import tseslint from 'typescript-eslint';
 
 const sourceFiles = ['**/*.{js,mjs,cjs,ts,mts,cts}'];
-const typescriptFiles = [
-  '**/src/**/*.{ts,mts,cts}',
-  '**/test/**/*.spec.{ts,mts,cts}',
-  '**/test/**/*.test.{ts,mts,cts}',
-  '**/test/**/__test__/**/*.{ts,mts,cts}',
-  '**/vitest/**/*.spec.{ts,mts,cts}',
-  '**/vitest/**/*.test.{ts,mts,cts}',
-];
-const jestTestFiles = [
-  '**/src/**/*.spec.{ts,mts,cts}',
-  '**/src/**/*.test.{ts,mts,cts}',
-  '**/src/**/__test__/**/*.{ts,mts,cts}',
-  '**/test/**/*.spec.{ts,mts,cts}',
-  '**/test/**/*.test.{ts,mts,cts}',
-  '**/test/**/__test__/**/*.{ts,mts,cts}',
-];
-const vitestTestFiles = ['**/vitest/**/*.spec.{ts,mts,cts}', '**/vitest/**/*.test.{ts,mts,cts}'];
+const typescriptFiles = ['**/src/**/*.{ts,mts,cts}', '**/test/**/*.{spec,test}.{ts,mts,cts}', '**/vitest/**/*.{spec,test}.{ts,mts,cts}'];
+const jestTestFiles = ['**/src/**/*.{spec,test}.{ts,mts,cts}', '**/test/**/*.{spec,test}.{ts,mts,cts}'];
+const vitestTestFiles = ['**/vitest/**/*.{spec,test}.{ts,mts,cts}'];
 const configDirname = path.dirname(url.fileURLToPath(import.meta.url));
 
 export default defineConfig([
   {
     name: 'Global Ignores',
-    ignores: ['**/.cache', '**/build', '**/coverage', '**/dist', '**/jest', '**/node_modules', '**/screenshots', '**/temp', '**/vendor', '**/apps', '**/chip'],
+    ignores: ['**/.cache', '**/apps', '**/build', '**/chip', '**/coverage', '**/dist', '**/jest', '**/node_modules', '**/screenshots', '**/temp', '**/vendor'],
   },
   {
     name: 'JavaScript & TypeScript Source Files',
@@ -104,7 +90,7 @@ export default defineConfig([
       parser: tseslint.parser,
       parserOptions: {
         tsconfigRootDir: configDirname,
-        project: './tsconfig.json',
+        project: 'tsconfig.json',
       },
     },
     // Comment out this line if you want to enable strict type-checked rules, but be aware that it may cause many errors until you fix all type issues in your codebase
@@ -144,7 +130,7 @@ export default defineConfig([
       parser: tseslint.parser,
       parserOptions: {
         tsconfigRootDir: configDirname,
-        project: './tsconfig.jest.json', // Use a separate tsconfig for Jest tests
+        project: 'tsconfig.jest.json', // Use a separate tsconfig for Jest tests
       },
     },
     extends: [jest.configs['flat/recommended']],
@@ -173,7 +159,7 @@ export default defineConfig([
       parser: tseslint.parser,
       parserOptions: {
         tsconfigRootDir: configDirname,
-        project: './tsconfig.vitest.json', // Use a separate tsconfig for Vitest tests
+        project: 'tsconfig.vitest.json', // Use a separate tsconfig for Vitest tests
       },
     },
     extends: [vitest.configs.recommended],
@@ -198,23 +184,21 @@ export default defineConfig([
   {
     name: 'JSON Files',
     files: ['**/*.json'],
-    ignores: ['**/devcontainer.json', '**/.vscode/*.json', '**/package-lock.json'],
+    ignores: ['**/devcontainer.json', '**/.vscode/*.json', '**/.oxlintrc.json', '**/.oxfmtrc.json', '**/package-lock.json'],
     plugins: { json, prettier },
     language: 'json/json',
     extends: [json.configs.recommended],
     rules: {
-      'json/no-unsafe-values': 'off',
       'prettier/prettier': 'warn', // Use Prettier for formatting
     },
   },
   {
     name: 'JSON with Comments Files',
-    files: ['**/*.jsonc', '**/devcontainer.json', '**/.vscode/*.json'],
+    files: ['**/*.jsonc', '**/devcontainer.json', '**/.vscode/*.json', '**/.oxlintrc.json', '**/.oxfmtrc.json'],
     plugins: { json, prettier },
     language: 'json/jsonc',
     extends: [json.configs.recommended],
     rules: {
-      'json/no-unsafe-values': 'off',
       'prettier/prettier': 'warn', // Use Prettier for formatting
     },
   },